organizing and md5 file hash
