+++ errors.go
@@ -1,0 +1,7 @@
+package helpers
+
+func check(e error) {
+	if e != nil {
+		panic(e)
+	}
+}
